Use relative time in action log.
